+++ public/data/mainFeedData.json
@@ -1,0 +1,26 @@
+[
+  {
+    "id": 1,
+    "userId": "navi",
+    "img": "./images/jiniChoi/feed1.png",
+    "likes": "sagong님 외 10명이 좋아합니다."
+  },
+  {
+    "id": 2,
+    "userId": "biki",
+    "img": "./images/jiniChoi/feed1.png",
+    "likes": "jini님 외 1087명이 좋아합니다."
+  },
+  {
+    "id": 3,
+    "userId": "gilyong",
+    "img": "./images/jiniChoi/feed1.png",
+    "likes": "dongin님 외 10234827명이 좋아합니다."
+  },
+  {
+    "id": 4,
+    "userId": "moa",
+    "img": "./images/jiniChoi/feed1.png",
+    "likes": "wecode님 외 2명이 좋아합니다."
+  }
+]
