+++ src/pages/jiniChoi/Comment/CommentBox.js
@@ -1,22 +1,50 @@
-import React from 'react';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faHeart } from '@fortawesome/free-solid-svg-icons';
+import React, { useState } from 'react';
+import Comment from '../Comment/CommentList';
+// import './CommentList.scss';
 
-const Comment = ({ newComment }) => {
+const CommentBox = () => {
+  const [commentList, setCommentList] = useState([]);
+  const [newComment, setNewComment] = useState('');
+
+  const handleCommentSubmit = e => {
+    // 1. 어떤 feed인지 알아야 함 (인자로 받는다)
+    // 2. 해당 feed의 commentsList 배열에 해당 리뷰를 추가한 형태로 setFeeds
+    e.preventDefault();
+    if (newComment.trim() === '') {
+      return;
+    }
+
+    setCommentList([...commentList, newComment]);
+    setNewComment('');
+  };
+
+  const newCommentInput = event => setNewComment(event.target.value);
+
   return (
-    <li className="commentBox">
-      <div className="left">
-        <div className="commentId">genie</div>
-        <div className="comment">{newComment}</div>
+    <>
+      <div className="commentsBox">
+        <ul>
+          {commentList.map((newComment, index) => (
+            <Comment newComment={newComment} key={index} />
+          ))}
+        </ul>
       </div>
 
-      <div className="right">
-        <button className="commentRemoveBtn">X</button>
-        <button className="commentlikeBtn">
-          <FontAwesomeIcon icon={faHeart} className="commentlike" />
+      <form className="commentPost" onSubmit={handleCommentSubmit}>
+        <input
+          required
+          type="text"
+          placeholder="댓글 달기..."
+          className="inputText"
+          value={newComment}
+          onChange={newCommentInput}
+        />
+        <button type="submit" className="submitBtn">
+          게시
         </button>
-      </div>
-    </li>
+      </form>
+    </>
   );
 };
-export default Comment;
+
+export default CommentBox;
