+++ src/pages/jiniChoi/Comment/CommentBox.scss
@@ -1,36 +1,43 @@
-li.commentBox {
+@import '../Comment/CommentList';
+
+.commentsBox {
+  margin: 5px 15px;
+
+  ul {
+    display: flex;
+    flex-direction: column;
+
+    // li.commentList
+  }
+}
+
+.commentPost {
   display: flex;
   justify-content: space-between;
-  line-height: 20px;
+  height: 40px;
+  padding: 0 15px;
+  border-top: 1px solid $BoxborderDefaultColor;
 
-  .left {
-    display: flex;
-    .commentId {
-      margin-right: 10px;
-      font-weight: bold;
+  .inputText {
+    width: 500px;
+    border: none;
+
+    &::placeholder {
+      color: $InputDefaultColor;
+    }
+    &:focus {
+      outline: none;
     }
   }
+  .submitBtn {
+    border: none;
+    background: none;
+    color: #55b6f9;
+    font-weight: bold;
+    cursor: pointer;
 
-  .right {
-    display: flex;
-
-    .commentRemoveBtn {
-      border: none;
-      background: none;
-      color: $subText;
-      cursor: pointer;
-    }
-    .commentlikeBtn {
-      margin-left: 5px;
-      border: none;
-      background: none;
-      /* color: $subText; */
-      color: #ee4956;
-      font-size: 15px;
-      cursor: pointer;
-    }
-    .commentlikeBtn.active {
-      color: #ee4956;
+    &:focus {
+      outline: none;
     }
   }
 }
